fix: Deprecate and remove old patch
