+++ patches/api/patches.api
@@ -56,13 +56,13 @@ public final class app/revanced/patches/all/misc/debugging/EnableAndroidDebuggin
 	public static final fun getEnableAndroidDebuggingPatch ()Lapp/revanced/patcher/patch/ResourcePatch;
 }
 
+public final class app/revanced/patches/all/misc/directory/ChangeDataDirectoryLocationPatchKt {
+	public static final fun getChangeDataDirectoryLocationPatch ()Lapp/revanced/patcher/patch/BytecodePatch;
+}
+
 public final class app/revanced/patches/all/misc/directory/documentsprovider/AddInternalDataDocumentsProviderKt {
 	public static final fun getInternalDataDocumentsProviderExtensionPatch ()Lapp/revanced/patcher/patch/BytecodePatch;
 	public static final fun getInternalDataDocumentsProviderPatch ()Lapp/revanced/patcher/patch/ResourcePatch;
-}
-
-public final class app/revanced/patches/all/misc/directory/redirect/ChangeDataDirectoryLocationPatchKt {
-	public static final fun getChangeDataDirectoryLocationPatch ()Lapp/revanced/patcher/patch/BytecodePatch;
 }
 
 public final class app/revanced/patches/all/misc/hex/HexPatchKt {
